"Deep" parser
For every possible token, add an attribute, 'next_context', which tells
the parser which set of tokenization rules should be used for the
following token. For example, if the lexer finds a token that begins a
string, the parser should use intra-string lexing rules until the token
representing the end of the string is found.