I still don't know what I'm doing.
kent.pl now only interfaces with the parser. The parser will create,
destroy, and use lexers as it sees fit. None of this is actually
implemented, mind you.

Newline logic is also smart enough now that the "width" part of the rule
table is unnecessary; therefore, remove the width column.