new approach to parser generation
Since the grammar no longer has any meta-qualities, it can be expressed
in a spreadsheet (.csv file in this case. Maybe that should change?)
We can either generate parser code from the grammar, or have generalized
code that just parses based on whatever's in the data structure
generated from the spreadsheet.